python/scrape_tool/src/requirements.txt
scrapy
python-dotenv
